cleanup, fix missing duration
